requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
proglog==0.1.9
pyee==9.0.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
